React Course Directory - React Router 4 App
*** React Course Directory - React Router 4 App ***

The application renders several routes using React Router to create a
SPA.

Dynamically inserted URLs are provided if a user wishes to provide them
on the Home route. The URL is updated using the push method on the
history object.

Routes manage the navigation (and sub-navigation) of the application
and Switch component aids in the rendering of a 404 style page if an
entered URL does not match a path in the application.

Active links are styled and made active when path matches URL.